back/requirements.txt
alembic==1.7.5
appnope==0.1.2
asttokens==2.0.5
backcall==0.2.0
black==21.12b0
click==8.0.3
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.3
et-xmlfile==1.1.0
executing==0.8.2
Flask==2.0.2
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
ipykernel==6.7.0
ipython==8.0.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jupyter-client==7.1.2
jupyter-core==4.9.1
flask-jwt-extended=4.5.3
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.1
openpyxl==3.0.9
pandas==1.3.5
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.24
psycopg2==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.1
Pygments==2.11.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyzmq==22.3.0
six==1.16.0
SQLAlchemy==1.4.31
stack-data==0.1.4
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing-extensions==4.0.1
wcwidth==0.2.5
Werkzeug==2.0.2
